migrated to TypeScript and fixed all type errors htbkoo/react-codepen-embed
